requirements.txt
numpy
scipy
matplotlib
joblib
tqdm
pyyaml
colorlog
tqdm

opencv-python
scikit-video

torch==1.2.0
torchvision

protobuf==3.9.1
tensorboard
tensorboardX
moviepy

flake8
isort
black
mypy
typeshed

# face_recognition
